requirements.txt
PyAudio
numpy
matplotlib
scipy
joblib
pillow
scikit-learn